+++ finagle-netty4/src/main/scala/com/twitter/finagle/netty4/transport/ChannelTransportContext.scala
@@ -2,7 +2,6 @@ package com.twitter.finagle.netty4.transport
 
 import com.twitter.finagle.Status
 import com.twitter.finagle.transport.TransportContext
-import com.twitter.util.{Future, Promise, Time}
 import io.netty.channel.Channel
 import io.netty.handler.ssl.SslHandler
 import java.net.SocketAddress
@@ -21,14 +20,10 @@ private[finagle] final class ChannelTransportContext(val ch: Channel)
 
   // Accessible by the `ChannelTransport` and for testing.
   private[transport] val failed = new AtomicBoolean(false)
-  private[transport] val closed = new Promise[Throwable]
-  private[transport] val alreadyClosed = new AtomicBoolean(false)
 
   def status: Status =
     if (failed.get || !ch.isOpen) Status.Closed
     else Status.Open
-
-  def onClose: Future[Throwable] = closed
 
   def localAddress: SocketAddress = ch.localAddress
 
@@ -42,14 +37,6 @@ private[finagle] final class ChannelTransportContext(val ch: Channel)
       } catch {
         case NonFatal(_) => None
       }
-  }
-
-  def close(deadline: Time): Future[Unit] = {
-    // we check if this has already been closed because of a netty bug
-    // https://github.com/netty/netty/issues/7638.  Remove this work-around once
-    // it's fixed.
-    if (alreadyClosed.compareAndSet(false, true) && ch.isOpen) ch.close()
-    closed.unit
   }
 
   def executor: Executor = ch.eventLoop